add functions

+++ FRONT/src/components/Card/Card.jsx
@@ -4,6 +4,8 @@ import { useNavigate } from 'react-router-dom';
 import { AppContext } from '../../provider/appContext';
 import Service from '../../services/index';
 import Swal from 'sweetalert2'
+import Button from '../Button/Button';
+import Modal from '../Modal/Modal';
 
 const Card = ({movie}) => {
 
@@ -16,17 +18,40 @@ const Card = ({movie}) => {
     }
 
     const deleteMovie = (id) => {
-        service.deleteData('movies', id)
-            .then(response => {
-                Swal.fire({
-                    icon: 'success',
-                    title: 'Movie added!',
-                    text: 'Movie added successfully!',
+
+        Swal.fire({
+            title: 'Are you sure?',
+            showDenyButton: true,
+            /* showCancelButton: true, */
+            confirmButtonText: 'Yes',
+          }).then((result) => {
+            /* Read more about isConfirmed, isDenied below */
+            if (result.isConfirmed) {
+                service.deleteData('movies', id)
+                .then(response => {
+                    Swal.fire({
+                        icon: 'success',
+                        title: 'Movie deleted',
+                        text: 'The movie has been deleted!',
+                    })
+                    getItems();
                 })
-                getItems();
-            })
+            } else if (result.isDenied) {
+              Swal.fire('Changes are not saved', '', 'info')
+            }
+          })
+
+
+
+        
     }
 
+    const editMovie = (movie) => {
+        
+        return (
+            <Modal title={'Update Movie From Database' + movie.name} type={'add'} id={'update_movie'}  movie={movie} />
+        )
+    }
 
 
 
@@ -34,33 +59,44 @@ const Card = ({movie}) => {
         <div className='col-12 col-md-6 col-lg-4 col-xl-3 mb-2'>
             <div className='d-flex flex-column align-items-center'>
                 <div className="card mt-4">
-                    <div className="card2">
+                    <div className="card2" onClick={() => showMore(movie.id)}>
                         <img
                             src={movie.image}
                             alt="iterstellar"
                         />
                     </div>
                 </div>
-                <h3 className='text-white text-center'>{movie.name}</h3>
+                <h3 className='text-white text-center mt-2'>{movie.name}</h3>
                 <div className='d-flex'>
-                    <button type="button" className="btn btn-primary me-1" onClick={() => {
-                        showMore(movie.id)
-                    }}>
-                        More
-                    </button>
+                    
                     {
                         srole === 'admin' &&
                         <>
-                        <button type="button" className="btn btn-danger me-1" onClick={() => {
+                        {/* <button data-bs-toggle="modal" data-bs-target="#update_movie" type="button" className="btn btn-warning  me-2">
+                            Update
+                        </button> */}
+
+                        <button type="button" className="btn btn-warning me-4" data-bs-toggle="modal" data-bs-target={`#modal_${movie.id}`} /* onClick={() => {
+                                editMovie(movie)
+                            }} */
+                            >
+                                Update
+                        </button>
+
+                        <button type="button" className="btn btn-danger" onClick={() => {
                             deleteMovie(movie.id)
                         }}>
                             Delete
                         </button>
-                        <button type="button" className="btn btn-warning" onClick={() => {
-                            deleteMovie(movie.id)
-                        }}>
-                            Edit
-                        </button>
+                        <div>
+                        
+                        </div>
+                            {/* <button type="button" className="btn btn-primary" data-bs-toggle="modal" data-bs-target={`#modal_${movie.id}`} 
+                        >
+                                Show More
+                        </button> */}
+                        
+                        <Modal title={`Update Movie From Database ${movie.name}`} type={'update'} id={`modal_${movie.id}`} movie={movie}  />
                         </>
 
                     }
